Split berny.tests into framework and per-potential modules
Move the ModelPotential base and run_and_check driver into berny.tests.base
and each model potential into its own module (berny.tests.linear_bend), so new
potentials get their own files. Public names are re-exported from the package
root, so `from berny.tests import LinearBendCrossover, run_and_check` is
unchanged.
